+++ src/event/SocketEvent.hxx
@@ -15,14 +15,14 @@ class SocketEvent {
 
     Event event;
 
-    typedef BoundMethod<void(short events)> Callback;
+    typedef BoundMethod<void(unsigned events)> Callback;
     const Callback callback;
 
 public:
     SocketEvent(EventLoop &_event_loop, Callback _callback)
         :event_loop(_event_loop), callback(_callback) {}
 
-    SocketEvent(EventLoop &_event_loop, evutil_socket_t fd, short events,
+    SocketEvent(EventLoop &_event_loop, evutil_socket_t fd, unsigned events,
                 Callback _callback)
         :SocketEvent(_event_loop, _callback) {
         Set(fd, events);
@@ -38,11 +38,11 @@ public:
     }
 
     gcc_pure
-    short GetEvents() const {
+    unsigned GetEvents() const {
         return event.GetEvents();
     }
 
-    void Set(evutil_socket_t fd, short events) {
+    void Set(evutil_socket_t fd, unsigned events) {
         event.Set(event_loop, fd, events, EventCallback, this);
     }
 
@@ -59,7 +59,7 @@ public:
     }
 
     gcc_pure
-    bool IsPending(short events) const {
+    bool IsPending(unsigned events) const {
         return event.IsPending(events);
     }
