+++ src/pg/AsyncConnection.cxx
@@ -201,7 +201,7 @@ AsyncPgConnection::ScheduleReconnect()
 }
 
 inline void
-AsyncPgConnection::OnSocketEvent(short)
+AsyncPgConnection::OnSocketEvent(unsigned)
 {
     switch (state) {
     case State::UNINITIALIZED:
